refactor(utils): share internal URL parsing
