+++ packages/coding-agent/src/internal-urls/index.ts
@@ -17,7 +17,6 @@ export * from "./local-protocol";
 export * from "./mcp-protocol";
 export * from "./memory-protocol";
 export * from "./omp-protocol";
-export * from "./parse";
 export * from "./router";
 export * from "./rule-protocol";
 export * from "./security-protocol";
