+++ packages/coding-agent/test/internal-urls/vault-protocol.test.ts
@@ -4,13 +4,12 @@ import * as os from "node:os";
 import * as path from "node:path";
 import {
 	InternalUrlRouter,
-	parseInternalUrl,
 	parseVaultUrl,
 	resolveVaultUrlToPath,
 	VaultProtocolHandler,
 } from "@oh-my-pi/pi-coding-agent/internal-urls";
 import * as vaultProtocol from "@oh-my-pi/pi-coding-agent/internal-urls/vault-protocol";
-import { removeWithRetries } from "@oh-my-pi/pi-utils";
+import { parseInternalUrl, removeWithRetries } from "@oh-my-pi/pi-utils";
 
 async function withTempDir<T>(fn: (dir: string) => Promise<T>): Promise<T> {
 	const dir = await fs.mkdtemp(path.join(os.tmpdir(), "vault-protocol-"));
