+++ packages/coding-agent/test/marketplace/parse-internal-url.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "bun:test";
-import { parseInternalUrl } from "@oh-my-pi/pi-coding-agent/internal-urls/parse";
+import { parseInternalUrl } from "@oh-my-pi/pi-utils";
 
 // ── Basic parsing (URLs that new URL() handles fine) ─────────────────
