+++ packages/coding-agent/test/tools/split-internal-url-sel.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, it } from "bun:test";
-import { pathTargetsSsh, peelWriteUrlSelector, splitInternalUrlSel } from "@oh-my-pi/pi-coding-agent/tools/path-utils";
+import { pathTargetsSsh, peelWriteUrlSelector } from "@oh-my-pi/pi-coding-agent/tools/path-utils";
+import { splitInternalUrlSel } from "@oh-my-pi/pi-utils";
 
 describe("splitInternalUrlSel", () => {
 	it("returns the input unchanged when there is no selector tail", () => {
